Extract css class functions in own file
